+++ src/test/java/org/logic2j/core/impl/DefaultTermMarshallerTest.java
@@ -61,7 +61,7 @@ public class DefaultTermMarshallerTest extends PrologTestBase {
     public void xBoundToY() {
         Object term = getProlog().getTermUnmarshaller().unmarshall("f(X, Y)");
         Bindings bindings = new Bindings(term);
-        bindings.getBinding((short) 0).bindTo(bindings.getBinding((short) 1).getVar(), bindings);
+        bindings.getBinding(0).bindTo(bindings.getBinding(1).getVar(), bindings);
         CharSequence formatted = getProlog().getTermMarshaller().marshall(term);
         assertEquals("f(X, Y)", formatted);
     }
@@ -70,7 +70,7 @@ public class DefaultTermMarshallerTest extends PrologTestBase {
     public void yBoundToX() {
         Object term = getProlog().getTermUnmarshaller().unmarshall("f(X, Y)");
         Bindings bindings = new Bindings(term);
-        bindings.getBinding((short) 1).bindTo(bindings.getBinding((short) 0).getVar(), bindings);
+        bindings.getBinding(1).bindTo(bindings.getBinding(0).getVar(), bindings);
         CharSequence formatted = getProlog().getTermMarshaller().marshall(term);
         assertEquals("f(X, Y)", formatted);
     }
